channel: Move Asset type and add more checks regarding the native asset for channel state conversion
