+++ channel/types/asset.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package channel
+package types
 
 import (
 	"perun.network/go-perun/channel"
